+++ backend/sister_site/src/main/java/domaine/user/UserDTO.java
@@ -50,8 +50,8 @@ public interface UserDTO extends BasicUserDTO {
   void fullFillUser(int id, String username, String lastName, String firstName, String email,
       boolean isBoss, Timestamp registrationDate, String password, int profilePicture, int address,
       String phoneNumber, String facebook, String instagram, String twitter, String youtube,
-      int hairColor, int hairSize, int eye, int height, int weight, int firstNationality,
-      int secondNationality, int shoeSize, int jacketSize, int pantSize, int chest, char braCup,
-      int waistSize, int hipSize, int neckSize, int headSize);
+      int hairColor, int hairSize, int eye, double height, int weight, int firstNationality,
+      int secondNationality, int shoeSize, int jacketSize, int pantSize, double chest, char braCup,
+      double waistSize, double hipSize, double neckSize, double headSize);
 
 }
